Update admin authority

+++ app/Http/Controllers/UsersController.php
@@ -18,15 +18,25 @@ class UsersController extends Controller
      */
     public function index()
     {
-        $users = User::all();
+        $emailActive = auth()->user()->email;
+        if($emailActive == 'admin@admin.com'){
+            $users = User::all();
+        }
+        else if($emailActive != 'admin@admin.com'){
+            $users = User::where('email', $emailActive)->get();
+        }
         return view('administrators.users.index', compact('users'));
     }
 
     public function archive()
     {
-        // $users = User::onlyTrashed()->get();
-        $users = User::onlyTrashed()->get();
-        return view('administrators.users.trash', compact('users'));
+        if(auth()->user()->email == 'admin@admin.com'){
+            $users = User::onlyTrashed()->get();
+            return view('administrators.users.trash', compact('users'));
+        }
+        else if(auth()->user()->email != 'admin@admin.com'){
+            return redirect()->back();
+        }
     }
     /**
      * Show the form for creating a new resource.
@@ -35,7 +45,13 @@ class UsersController extends Controller
      */
     public function create()
     {
-        return view('administrators.users.create');
+        $emailActive = auth()->user()->email;
+        if($emailActive == 'admin@admin.com'){
+            return view('administrators.users.create');
+        }
+        else if($emailActive != 'admin@admin.com'){
+            return redirect()->back()->with('failed', 'You do not have the authority to access this feature!');
+        }
     }
 
     /**
@@ -94,19 +110,33 @@ class UsersController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-            'name' => 'required',
-            'email' => 'required',
-            'password' => 'required',
-        ]);
-
-        User::where('id', $id)
-        ->update([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password)
-        ]);
-
+        $user = User::where('id', $id)->get();
+        if($user[0]->email != "admin@admin.com"){
+            $request->validate([
+                'name' => 'required',
+                'email' => 'required',
+                'password' => 'required',
+            ]);
+    
+            User::where('id', $id)
+            ->update([
+                'name' => $request->name,
+                'email' => $request->email,
+                'password' => Hash::make($request->password)
+            ]);
+        }
+        else if($user[0]->email == "admin@admin.com"){
+            $request->validate([
+                'name' => 'required',
+                'password' => 'required',
+            ]);
+    
+            User::where('id', $id)
+            ->update([
+                'name' => $request->name,
+                'password' => Hash::make($request->password)
+            ]);
+        }
         return redirect()->back()->with('success', 'User updated!');
     }
 
@@ -118,19 +148,37 @@ class UsersController extends Controller
      */
     public function destroy($id)
     {
-        User::where('id', $id)->delete();
-        return redirect()->back()->with('success', 'User archived!');
+        $emailActive = auth()->user()->email;
+        if($emailActive == 'admin@admin.com'){
+            User::where('id', $id)->delete();
+            return redirect()->back()->with('success', 'User archived!');
+        }
+        else if($emailActive == 'admin@admin.com'){
+            return redirect()->back()->with('failed', 'You do not have the authority to access this feature!');
+        }
     }
 
     public function kill($id)
     {
-        User::onlyTrashed()->where('id', $id)->forceDelete();
-        return redirect()->back()->with('success', 'User deleted!');
+        $emailActive = auth()->user()->email;
+        if($emailActive == 'admin@admin.com'){
+            User::onlyTrashed()->where('id', $id)->forceDelete();
+            return redirect()->back()->with('success', 'User deleted!');
+        }
+        else if($emailActive == 'admin@admin.com'){
+            return redirect()->back()->with('failed', 'You do not have the authority to access this feature!');
+        }
     }
 
     public function restore($id)
     {
-        User::onlyTrashed()->where('id', $id)->restore();
-        return redirect()->back()->with('success', 'User restored!');
+        $emailActive = auth()->user()->email;
+        if($emailActive == 'admin@admin.com'){
+            User::onlyTrashed()->where('id', $id)->restore();
+            return redirect()->back()->with('success', 'User restored!');
+        }
+        else if($emailActive == 'admin@admin.com'){
+            return redirect()->back()->with('failed', 'You do not have the authority to access this feature!');
+        }
     }
 }
